+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## Unreleased
+
+
+## 0.10.1 - 2021-11-02
 * `now_ns` now returns nanoseconds since unix epoch.
 * Make scope merging deterministic.